Adding riemann_solve_strong_shock_tube.
This is an example Riemann solver for a single test problem.  This agrees with
some fortran code that I found online, although in our case, we get many more
digits of accuracy.

This routine needs to be massively cleaned up.  We need a bit of cleaner data
structure to contain all of the available options that might be included.