Changed name of Riemann sovler for Euler.
This now has a much nicer user friendly output.

The callable functions representing the rarefaction wave still need to be
checked, and it would be nice to glue together each region.  Ideally, we'd
like this to be a single callable function that presents a list of point
values of the form, [x,y] that can be plotted.