Burger's equation.
Need to add q_init_partial_x to DefineTestCase class.  This function needed to
find the exact solution.

Adding sine_to_n.py which computes the exact solution.   This function doesn't
assume that q is a list of np arrays, so in order to use this, we need a single
wrapper function to index into the list.

It would be nice to glue this into a batch_refinement script for running a
convergence study.